requirements.txt
django
mysqlclient
djangorestframework
djoser
gunicorn
django-heroku
boto3
django-rest-knox